+++ R/utils.R
@@ -30,15 +30,24 @@ ftime = function(st, digits = 3) {
   unique.default(x[match(x, y, 0L) == 0L])
 }
 
-pluralise = function(noun, n) {
-  if(n == 1) noun else sprintf("%ss", noun)
+pluralise = function(noun, n, numberFirst = TRUE) {
+  s = if(n == 1) noun else sprintf("%ss", noun)
+  if(numberFirst) 
+    s = paste(n, s)
+  s
 }
 
+.safeMax = function(v) {
+  if(length(v) == 0 || all(is.na(v))) 
+    return(NA)
+  max(v, na.rm = TRUE)
+}
 
-trunc = function(x, printMax = 10) {
+trunc = function(x, printMax = 8, head = 3) {
   if(length(x) <= printMax)
     return(toString(x))
-  y = c(x[1:5], "...", x[length(x)])
+  maxHead = max(1, min(head, printMax - 2))
+  y = c(head(x, maxHead), "...", x[length(x)])
   toString(y)
 }
 
@@ -73,10 +82,16 @@ sprintfNamed = function(fmt, ...) {
 }
 
 # Undisputed entries in a LR/GLR matrix
-undisputedEntries = function(M, threshold = 1, strict = TRUE) {
+undisputedEntries = function(M, threshold, strict = TRUE) {
+  
+  if(!length(M))
+    return(matrix(nrow = 0, ncol = 2, dimnames = list(NULL, c("row", "col"))))
+  
+  # Replace NAs with -1 to allow numeric comparisons below
+  M[is.na(M)] = -1
   
   # Indices of matches exceeding threshold
-  highIdx = which(M > threshold, arr.ind = TRUE)
+  highIdx = which(M >= threshold - sqrt(.Machine$double.eps), arr.ind = TRUE)
   
   # Return if empty
   if(!nrow(highIdx)) 
@@ -99,3 +114,5 @@ undisputedEntries = function(M, threshold = 1, strict = TRUE) {
   # Return matrix of indices of undisputed matches   
   highIdx[isUndisp, , drop = FALSE]
 }
+
+